refactor(commands): extract run.rs from commands/mod.rs (Issue #271)
Moves the run/exec cluster (run_script, run_python_code,
find_python_interpreter, script lockfile helpers, sandbox info, PEP 723
index settings) out of commands/mod.rs into a dedicated commands/run.rs
module, following the pattern established in PR #371 for install.rs.

This covers only the run.rs portion of Issue #271's proposed scope.
lock.rs and remaining extraction remain as follow-up work.

Progress on #271 — not closing yet.
